refactor(arc): extract sort-filters feature
