+++ src/features/abstract-resource-collection/types/index.ts
@@ -1,7 +1,1 @@
 export type { ListItem } from "./internal";
-
-export type {
-  FilteredField,
-  FilterDefinitions,
-  SortFiltersFormValuesNarrowed,
-} from "./sort-filters";
