+++ src/features/sort-filters/data/filter-type-mappings.ts
