+++ src/features/sort-filters/data/sort-filter-default-values.ts
@@ -1,5 +1,5 @@
 import { SortDirection } from "../enums";
-import type { SortFiltersFormValuesNarrowed } from "../types/sort-filters";
+import type { SortFiltersFormValuesNarrowed } from "../types/internal";
 
 export const SORT_FILTER_DEFAULT_VALUES = {
   sortBy: null,
