+++ src/features/sort-filters/types/index.ts
@@ -1,0 +1,5 @@
+export type {
+  FilteredField,
+  FilterDefinitions,
+  SortFiltersFormValuesNarrowed,
+} from "./internal";
