+++ src/features/sort-filters/utils/serialize-sort-filters.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from "vitest";
 
 import { SortDirection } from "../enums";
-import type { SortFiltersFormValuesNarrowed } from "../types/sort-filters";
+import type { SortFiltersFormValuesNarrowed } from "../types/internal";
 import { serializeSortFilters } from "./serialize-sort-filters";
 
 describe("serializeSortFilters function", () => {
